pyproject.toml
[project]
name = "haptgp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "adafruit-circuitpython-bme280>=2.6.29",
    "adafruit-circuitpython-color-terminal>=0.2.0",
    "adafruit-circuitpython-displayio-layout>=3.1.1",
    "adafruit-circuitpython-drv2605>=1.3.7",
    "adafruit-circuitpython-ds3231>=2.4.25",
    "adafruit-circuitpython-gc9a01a>=1.0.0",
    "adafruit-circuitpython-imageload>=1.24.4",
    "adafruit-circuitpython-neopixel>=6.3.17",
    "adafruit-circuitpython-veml7700>=2.1.5",
    "circuitpython-cst816>=0.1.0",
    "fastapi>=0.118.0",
    "gpiozero>=2.0.1",
    "matplotlib>=3.10.7",
    "paho-mqtt>=2.1.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "reaktiv>=0.19.2",
    "rpi-gpio>=0.7.1",
    "rpi-ws281x>=5.0.0",
    "scipy>=1.16.3",
    "smbus2>=0.5.0",
    "sshkeyboard>=2.3.1",
    "typed-argument-parser>=1.11.0",
]
